Refactoring : ModalFlow (#455)
Several components have 3-step modal flows for important actions :
1 - modal where the user confirms the action
2 - modal where user waits while action is happening in the backend
3 - success modal (or if error, back to 1 with error message displayed)

This PR adds a ModalFlow component to do this, and uses it for ControlDeleteFlow, QuestionnaireDeleteFlow and PublishFlow.
